cdk/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecs_patterns
aws-cdk.aws_elasticloadbalancingv2

# Work around for jsii#413
aws-cdk.aws-autoscaling-common

python-dotenv
